Создан класс GradientProgressBar, который обособляет логику отрисовки градиента
